Shifted Viz code to the HomeScreen, and resolved fbjs dependency
