+++ frontend/package.json
@@ -9,17 +9,18 @@
   },
   "dependencies": {
     "expo": "^35.0.0",
+    "fbjs": "^1.0.0",
     "lodash": "^4.17.15",
     "react": "16.8.3",
     "react-dom": "16.8.3",
     "react-native": "https://github.com/expo/react-native/archive/sdk-35.0.0.tar.gz",
+    "react-native-gesture-handler": "~1.3.0",
     "react-native-reanimated": "^1.4.0",
+    "react-native-svg": "^9.13.3",
     "react-native-web": "^0.11.7",
     "react-navigation": "^3.12.0",
     "react-navigation-stack": "^1.10.3",
     "react-navigation-tabs": "^2.5.6",
-    "react-native-gesture-handler": "~1.3.0",
-    "react-native-svg": "^9.13.3",
     "victory-native": "^33.0.0"
   },
   "devDependencies": {
