+++ frontend/screens/HomeScreen/index.js
@@ -1,21 +1,19 @@
 import React from 'react';
 import {StyleSheet, Text, View} from 'react-native';
 import {Button} from 'react-native';
-export default function HomeScreen(props) {
-  const { navigate } = props.navigation;
-  return (
-    <Button title="Camera" onPress={() => navigate('Camera')} />
-  );
-}
+import { VictoryLine, VictoryChart } from "victory-native";
 
-HomeScreen.navigationOptions = {
-  header: null,
-};
+const data = [
+  { quarter: 1, earnings: 13000 },
+  { quarter: 2, earnings: 16500 },
+  { quarter: 3, earnings: 14250 },
+  { quarter: 4, earnings: 19000 }
+];
 
 const styles = StyleSheet.create({
     container: {
       flex: 1,
-      backgroundColor: '#fff',
+      backgroundColor: "#f5fcff",
       alignItems: 'center',
       justifyContent: 'center',
     },
@@ -26,4 +24,22 @@ const styles = StyleSheet.create({
       textAlign: 'center',
     },
 });
-  
+
+
+export default function HomeScreen(props) {
+  const { navigate } = props.navigation;
+  return (
+    <View style={styles.container}>
+      <Button title="Camera" onPress={() => navigate('Camera')} />
+      <VictoryChart width={350}>
+        <VictoryLine data={data}
+        alignment = "middle"
+        x="quarter" y="earnings" />
+      </VictoryChart>
+    </View>
+  );
+}
+
+HomeScreen.navigationOptions = {
+  header: null,
+};
